Completely overhaul scroll logic
This took some amount of effort. Quite proud of little nomad.

With a lot of help from the Microsoft docs on terminal sequences[1] I've
overhauled the scroll logic. Instead of using a single offset in a
buffer that expands with the content inserted into it, the scrollable
area is now confined to a set scroll region within the buffer. Scrolling
now involves moving lines in the buffer.

This is the logic required by programs that rely on being able to scroll
lines in a subset of the viewport. vim is an example of such a program.
less on the other hand will happily work regardless of this because in
less, the entire viewport is scrolled. However, before this fix there
was a bug where using less would sometimes cause ghost lines to be
inserted that scrolled with the intended lines. This bug has now
disappeared.

Moving lines (or at least pointers to lines) around in memory is not
exactly the most efficient way of doing things, so performance is
slightly degraded. In future, it might be best to consider data
structures that are more optimised for these transitions.

I've been testing vim a lot in the process of making this fix, and I'm
very pleased with the result.

vim is now useable within nomad!

[1] https://docs.microsoft.com/en-us/windows/console/console-virtual-terminal-sequences