Add record/integration scripts
These allow for recording arbitrary interactions with the host shell
within a subterminal and testing to make sure that the current
implementation of subterminal preserves the same output for the recorded
input.

Input consists of output recorded from processes (record-test-integ.js)
whereby arbitrary user input was given to start programs and interact
with them.

The current test data has been tracked in logs/test-integ.log.

check-test-integ.js can be run to check that given the input stored
test-integ.log, the output of subterminal.js matches the output
previously recorded in test-integ.log.

If a bug is fixed that alters the expected state given a process output,
the changes will need to be committed to test-integ.log.